requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
numpy
opencv-python
python-multipart
PyYAML
pandas
matplotlib
scipy
pyyaml